Install R Packages from Multiple Sources

A unified installer for R packages from CRAN, GitHub, Bioconductor, or local source.

Usage

inst_pkg(
  pkg = NULL,
  source = c("CRAN", "GitHub", "Bioconductor", "Local"),
  path = NULL,
  ...
)

Arguments

pkg

Package name(s) or GitHub repo (e.g., "user/repo"). Not required for source = "local".

source

Source of package: "CRAN", "GitHub", "Bioconductor", "local". Case-insensitive, shorthand allowed.

path

Path to local package file (used when source = "local").

...

Additional arguments passed to install.packages(), devtools::install_github(), or BiocManager::install().

Value

NULL (invisibly)

Examples

inst_pkg("dplyr", source = "CRAN")
#> ℹ Package dplyr is already installed. Skipped.
inst_pkg("hadley/emo", source = "GitHub")
#> ℹ Installing from GitHub: hadley/emo
#> Using GitHub PAT from the git credential store.
#> Downloading GitHub repo hadley/emo@HEAD
#> cli      (3.6.4 -> 3.6.5) [CRAN]
#> magrittr (2.0.3 -> 2.0.4) [CRAN]
#> generics (0.1.3 -> 0.1.4) [CRAN]
#> purrr    (1.0.4 -> 1.1.0) [CRAN]
#> stringr  (1.5.1 -> 1.5.2) [CRAN]
#> Installing 5 packages: cli, magrittr, generics, purrr, stringr
#> Installing packages into 'C:/Users/84241/AppData/Local/Temp/RtmpqO7KJ2/temp_libpath283c6e8554f4'
#> (as 'lib' is unspecified)
#> 
#>   There is a binary version available but the source version is later:
#>          binary source needs_compilation
#> magrittr  2.0.3  2.0.4              TRUE
#> 
#> package 'cli' successfully unpacked and MD5 sums checked
#> package 'generics' successfully unpacked and MD5 sums checked
#> package 'purrr' successfully unpacked and MD5 sums checked
#> package 'stringr' successfully unpacked and MD5 sums checked
#> 
#> The downloaded binary packages are in
#> 	C:\Users\84241\AppData\Local\Temp\RtmpGKD5bk\downloaded_packages
#> installing the source package 'magrittr'
#> ── R CMD build ─────────────────────────────────────────────────────────────────
#> * checking for file 'C:\Users\84241\AppData\Local\Temp\RtmpGKD5bk\remotesa09421cd37d6\hadley-emo-3f03b11/DESCRIPTION' ... OK
#> * preparing 'emo':
#> * checking DESCRIPTION meta-information ... OK
#> * checking for LF line-endings in source and make files and shell scripts
#> * checking for empty or unneeded directories
#> * building 'emo_0.0.0.9000.tar.gz'
#> 
#> Installing package into 'C:/Users/84241/AppData/Local/Temp/RtmpqO7KJ2/temp_libpath283c6e8554f4'
#> (as 'lib' is unspecified)
#> ✔ Installation complete!
inst_pkg("scRNAseq", source = "Bioconductor")
#> ℹ Package scRNAseq is already installed. Skipped.
inst_pkg(source = "local", path = "mypackage.tar.gz")
#> Error in match.arg(source, c("CRAN", "GitHub", "Bioconductor", "Local")): 'arg' should be one of "CRAN", "GitHub", "Bioconductor", "Local"