Paste two strings with a space

%p% is a custom infix operator for string concatenation with a single space. It’s inspired by the readability of %>%, but meant for expressive text building.

Usage

lhs %p% rhs

Arguments

lhs

A character vector on the left-hand side.

rhs

A character vector on the right-hand side.

Value

A character vector, with elements from lhs and rhs pasted together using a space.

Examples

"Hello" %p% "world"
#> [1] "Hello world"
"Good" %p% "job"
#> [1] "Good job"